bugs fix on limit

+++ server/controllers/coinController.js
@@ -1,6 +1,8 @@
 import axios from 'axios';
 import Coin from '../models/Coins.js';
 import History from '../models/History.js';
+import NodeCache from 'node-cache';
+const cache = new NodeCache({ stdTTL: 60 }); // cache lasts 60 seconds
 
 const API_URL =
   'https://api.coingecko.com/api/v3/coins/markets?vs_currency=usd&order=market_cap_desc&per_page=10&page=1';
@@ -30,6 +32,13 @@ const API_URL =
 // };
 export const getCoins = async (req, res) => {
   try {
+    const cachedData = cache.get("coins");
+
+    if (cachedData) {
+      console.log("Serving from cache 🚀");
+      return res.json(cachedData);
+    }
+
     const { data } = await axios.get(API_URL);
     await Coin.deleteMany({});
 
@@ -45,6 +54,9 @@ export const getCoins = async (req, res) => {
 
     await Coin.insertMany(formatted);
 
+    cache.set("coins", formatted); // Save result to cache
+    console.log("Fetched from API & cached ✅");
+
     res.json(formatted);
   } catch (err) {
     console.error('Error fetching coins:', err.message || err);
@@ -55,6 +67,7 @@ export const getCoins = async (req, res) => {
     res.status(500).json({ error: 'Failed to fetch coins', details: err.message });
   }
 };
+
 
 
 export const postHistory = async (req, res) => {
